+++ jfxapps.boot/jfxapps.boot.api/src/main/java/com/gluonhq/jfxapps/boot/api/context/ProgressListener.java
@@ -1,5 +1,5 @@
 package com.gluonhq.jfxapps.boot.api.context;
 
 public interface ProgressListener {
-    void notifyProgress(double progress);
+    void notifyProgress(float progress);
 }
